+++ packages/terra-hookshot/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Removed all inline styles from Dev-site components and implemented the same styles using external css.
 
+### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
+
 5.7.0 - (April 24, 2019)
 ------------------
 ### Changed