+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Removed all inline styles from Dev-site components and implemented the same styles using external css.
 
+### changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
+
 5.2.0 - (May 21, 2019)
 ------------------
 ### Added